user can now customize the exponential backoff's parameters
